Functions to improve the country attribution in a c14_date_list

c14bazAAR provides two functions to check and improve the spatial attribution of the individual dates in a c14_date_list to a country.

c14bazAAR::fix_database_country_name() adds column country_thes with standardized country names. Most source databases come with a column country that contains a character name of the origin country for each date. Unfortunately the different source databases don't rely on a unified naming convention and therefore use various terms to represent the same country (for example: United Kingdom, Great Britain, GB, etc.). This function aims to standardize the country naming scheme. To achieve this, it compares the names to values in an external (countrycode::codelist) and an internal country_thesaurus. The latter relies on manual curation to catch semantic and spelling errors in the source databases. See inspect_lookup_country to trace the lookup decisions.

c14bazAAR::determine_country_by_coordinate() adds the column country_coord with standardized country attribution based on the coordinate information of the dates. Due to the inconsistencies in the country column in many c14 source databases it's often necessary to rely on the coordinate position (lat & lon) for country attribution information. Unfortunately not all source databases store coordinates.

determine_country_by_coordinate(x, suppress_spatial_warnings = TRUE)

# S3 method for default
determine_country_by_coordinate(x, suppress_spatial_warnings = TRUE)

# S3 method for c14_date_list
determine_country_by_coordinate(x, suppress_spatial_warnings = TRUE)

fix_database_country_name(
  x,
  thesaurus = c14bazAAR::country_thesaurus,
  codesets = c("country.name.de", "iso3c"),
  ...
)

# S3 method for default
fix_database_country_name(
  x,
  thesaurus = c14bazAAR::country_thesaurus,
  codesets = c("country.name.de", "iso3c"),
  ...
)

# S3 method for c14_date_list
fix_database_country_name(
  x,
  thesaurus = c14bazAAR::country_thesaurus,
  codesets = c("country.name.de", "iso3c"),
  ...
)

Arguments

x
an object of class c14_date_list

suppress_spatial_warnings
suppress some spatial data messages and warnings

thesaurus
data.frame with "correct" country names and "wrong" variants

codesets
which country codesets should be searched for in countrycode::codelist beyond country.name.en? See ?countrycode::codelist for more information

...
additional arguments are passed to stringdist::stringdist(). stringdist() is used for fuzzy string matching of the country names in countrycode::codelist

Value

an object of class c14_date_list with the additional columns country_thes, country_coord and/or country_final

Examples

library(magrittr)
example_c14_date_list %>%
  determine_country_by_coordinate()

#> Determining country by coordinates... 

#> 	Radiocarbon date list
#> 	dates: 9
#> 	sites: 4
#> 	countries: 5
#> 	uncalBP: 9000 ― 1000 
#> 
#> # A data frame: 9 × 20
#>   sourcedb method labnr c14age c14std c13val site   sitetype   feature period   
#>   <chr>    <chr>  <chr>  <int>  <int>  <dbl> <chr>  <chr>      <chr>   <chr>    
#> 1 A        Conv   lab-1   1000     20    -15 Site A Burial mo… Grave 1 NA       
#> 2 A        Conv   lab-2   2000     30    -20 Site A NA         Grave 2 Bronze A…
#> 3 A        AMS    lab-3   3000     40    -25 Site B settlement NA      Bronze A…
#> 4 B        NA     lab-4   4000     50    -15 Site B settlement House 5 Late Neo…
#> 5 B        AMS    lab-5   5000     50    -20 Site C settlement NA      NA       
#> 6 B        AMS    lab-6   6000     60    -25 Site C settlement Oven 4  Early Ne…
#> 7 C        AMS    lab-7   7000     70    -15 Site D Camp       Hut 2   Mesolith…
#> 8 C        AMS    lab-8   8000     80    -20 Site D Camp       Hut 2   Mesolith…
#> 9 C        AMS    lab-9   9000     90    -25 Site D Camp       Hut 3   Mesolith…
#> # … with 10 more variables: culture <chr>, material <chr>, species <chr>,
#> #   region <chr>, country <chr>, country_coord <chr>, lat <dbl>, lon <dbl>,
#> #   shortref <chr>, comment <chr>

example_c14_date_list %>%
  fix_database_country_name()

#> Standardizing country names... 

#> 	Radiocarbon date list
#> 	dates: 9
#> 	sites: 4
#> 	countries: 5
#> 	uncalBP: 9000 ― 1000 
#> 
#> # A data frame: 9 × 20
#>   sourcedb method labnr c14age c14std c13val site   sitetype   feature period   
#>   <chr>    <chr>  <chr>  <int>  <int>  <dbl> <chr>  <chr>      <chr>   <chr>    
#> 1 A        Conv   lab-1   1000     20    -15 Site A Burial mo… Grave 1 NA       
#> 2 A        Conv   lab-2   2000     30    -20 Site A NA         Grave 2 Bronze A…
#> 3 A        AMS    lab-3   3000     40    -25 Site B settlement NA      Bronze A…
#> 4 B        NA     lab-4   4000     50    -15 Site B settlement House 5 Late Neo…
#> 5 B        AMS    lab-5   5000     50    -20 Site C settlement NA      NA       
#> 6 B        AMS    lab-6   6000     60    -25 Site C settlement Oven 4  Early Ne…
#> 7 C        AMS    lab-7   7000     70    -15 Site D Camp       Hut 2   Mesolith…
#> 8 C        AMS    lab-8   8000     80    -20 Site D Camp       Hut 2   Mesolith…
#> 9 C        AMS    lab-9   9000     90    -25 Site D Camp       Hut 3   Mesolith…
#> # … with 10 more variables: culture <chr>, material <chr>, species <chr>,
#> #   region <chr>, country <chr>, country_thes <chr>, lat <dbl>, lon <dbl>,
#> #   shortref <chr>, comment <chr>